+++ index.html
@@ -89,7 +89,7 @@
 								<object
 									class="counter__label__number"
 									data-js="userQuestionCounter">
-									50
+									0
 								</object>
 								<span data-js="userQuestionLabel">Questions</span>
 							</section>
